Create number_of_cache_items
This C++ code is implementing a solution to a problem involving ranges and queries. Let's break down the code step by step to understand what it's doing:

The code starts by including the necessary header <bits/stdc++.h> and defining a typedef for the data type long long int as ll. It also includes the standard namespace (using namespace std;).

Two functions are defined:

gcd(ll a, ll b): Computes the greatest common divisor (GCD) of two numbers a and b using the Euclidean algorithm.
power(ll x, ll y, ll p): Calculates the modular exponentiation x^y % p efficiently using binary exponentiation.
The main function begins:

It disables synchronization between the C and C++ standard streams (ios_base::sync_with_stdio(false);) for faster input/output.
It defines variables n and q to store the number of ranges and the number of queries, respectively. It reads the value of n from the input.
A map mp is used to store the frequency of events occurring at different points on a number line. The loop reads pairs of integers x and y representing the start and end points of a range. It increments the frequency of x in the map and decrements the frequency of x+y+1 in the map. This is done to keep track of the frequency of events occurring in each range.

The code then calculates the cumulative frequency of events using mp2. It iterates through the mp map, adding the current frequency to a cumulative sum pre and storing this cumulative sum in the mp2 map for each key (point on the number line).

The code reads the value of q (number of queries) from the input.

For each query, the code reads an integer x and performs the following steps:

It searches for the first element in the mp2 map that is greater than x using upper_bound.
If such an element is found, it backtracks one step using --it to get the element before the found element, then prints the second value of the key-value pair (which represents the cumulative frequency of events up to that point).
If no such element is found (i.e., all events occurred after x), it prints "0".
The main function returns 0, indicating successful program execution.

In summary, this code is designed to efficiently handle queries related to ranges and events occurring on a number line. It calculates cumulative frequencies of events and responds to queries asking for the number of events that occurred before a given point on the number line.