Create bs_rotatedarray
The concept here is amazing, because we dont have a sorted array but we knew it was originally sorted but now we rotated it. best part we can still apply binary search. We check either left or right half is sorted and the other half is not sorted. And then we apply bs on sorted part if we can find target there. If not then we check on unsorted array. Find middle element and then check which half of this unsorted half is sorted, and then apply bs there.